WIP: extract re_class from re_ast
